Small fixes with registration, customized my toolbar, added some description and features for my website on home page
Realized to NOT put spacing when registering a new username (for the user's actual username). Added some links in my dropdown box to go to main map page and home page, as well as sources from where I saw the job posting. It may be a little too ambitious for me to filter jobs by title and county but am going to attempt this.