Implemented a running form that works when a user is logged in to the main page
the guts of the project are done except for getting the added job onto my main map after the form is submitted. I was initially going to geocode an address but it became too problematic so I am back to the user adding it on a map (part of form) submitting the form with information of job and the job should be able to be viewable on my main map.